pyproject.toml
[tool.poetry]
name = "basic-reflexion-agent"
version = "0.1.0"
description = ""
authors = ["Brais Maneiro Sánchez <brais.maneiro@netexlearning.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchain-openai = "^0.1.20"
langgraph = "^0.1.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
